fixed type casting

+++ ngx_rtmp_cmd_module.c
@@ -288,8 +288,8 @@ ngx_rtmp_cmd_connect(ngx_rtmp_session_t *s, ngx_rtmp_connect_t *v)
         s->app.len = (p - s->app.data);
     }
 
-    s->acodecs = v->acodecs;
-    s->vcodecs = v->vcodecs;
+    s->acodecs = (uint32_t) v->acodecs;
+    s->vcodecs = (uint32_t) v->vcodecs;
 
     /* find application & set app_conf */
     cacfp = cscf->applications.elts;
